+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Refactor fields renderer to support custom actions on fields
 - Update Vue to version 3.3.7
+- Sort content models alphabetically on overview page
 
 ## 2.7.1 (2023-10-08)